model structure for NN
